% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_names.R
\name{fname_stamped}
\alias{fname_stamped}
\title{Create a time stamped string to use to save files}
\usage{
fname_stamped(name_base, name_extension)
}
\arguments{
\item{name_base}{the base of the file name}

\item{name_extension}{the file extension}
}
\description{
Create a time stamped string to use to save files
}
\examples{
fname_stamped("data",".csv")
}
