% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{read.excel}
\alias{read.excel}
\title{Read Excel cells to a dataframe}
\usage{
read.excel(header = TRUE, ...)
}
\arguments{
\item{header}{copy headers?}

\item{...}{}
}
\description{
Read Excel cells to a dataframe
}
\examples{
read.excel(df, header = TRUE)
}
