% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.r
\name{remove_na_columns}
\alias{remove_na_columns}
\title{Remove NA columns}
\usage{
remove_na_columns(df)
}
\arguments{
\item{df}{}
}
\value{
a data frame that only includes columns with at least one value
}
\description{
Remove NA columns
}
\examples{
remove_na_columns(df)
}
