% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.r
\name{remove_no_variance_columns}
\alias{remove_no_variance_columns}
\title{Remove columns with no variance}
\usage{
remove_no_variance_columns(df)
}
\arguments{
\item{df}{}
}
\value{
a data frame where zero variance columns are removed
}
\description{
Remove columns with no variance
}
\examples{
remove_no_variance_columns(df)
}
