% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{write.excel}
\alias{write.excel}
\title{Write dataframe to clipboard for pasting to Excel}
\usage{
write.excel(y, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{y}{a dataframe}

\item{row.names}{Include row names?}

\item{col.names}{Include col names?}

\item{...}{}
}
\description{
Write dataframe to clipboard for pasting to Excel
}
\examples{
write.excel(df, row.names = FALSE, col.names = TRUE)
}
